# openapi.yaml
openapi: 3.0.3
info:
  title: OOH Audio Sync API
  description: Backend APIs for the DreamBricks Audio-Video Sync Totem system.
  version: "1.0.0"
servers:
  - url: http://localhost:8001
    description: Local Node.js Server
paths:
...
  /api/login:
    post:
      summary: Admin Login
      description: |
        Checks `ADMIN_USER` / `ADMIN_PASSWORD` from `.env` and sets the `db_admin` HttpOnly cookie
        (a random session id; only its hash is stored, in `sessions.json` — in the bucket in S3 mode,
        shared by every server). The session lasts until logout. All other `/api/*` routes require this cookie.
        After 5 failed attempts the IP is locked for 60 seconds.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Logged in (cookie set)
        '401':
          description: Invalid username or password
        '429':
          description: Too many failed attempts
        '503':
          description: Login not configured in `.env`
  /api/logout:
    post:
      summary: Admin Logout
      description: Ends the session on the server (a copied cookie stops working) and clears the cookie.
      responses:
        '200':
          description: Logged out
        '502':
          description: The session could not be removed from the storage
...
  /api/sessions/revoke-others:
    post:
      summary: Disconnect Other Devices
      description: Ends every admin session except the one making the request.
      responses:
        '200':
          description: Sessions ended
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  revoked: { type: integer, example: 1 }
        '502':
          description: The sessions could not be saved to the storage
...